requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
flask-limiter==3.7.0
APScheduler==3.10.4
httpx==0.27.0
itsdangerous==2.2.0
Werkzeug==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.3.0
pywebpush==2.0.0
cryptography==42.0.8
Flask-Mail==0.10.0
